Implement remote DNS
This commit implements the ability to resolve IP addresses remotely via
SOCKS5 or HTTP. It does so by intercepting DNS A and AAAA queries and
replying with IP addresses from a virtual IP space which is reserved
for DNS resolves. When an application performs a DNS lookup, a mapping
from the virtual IP to a DNS name is created and cached for a short
period of time. If the application subsequently connects to the virtual
IP, the dialer will use the DNS name from the mapping instead of the
destination IP address.